Mysql/ConectMysql/Conexao/requirements.txt
# Libs
mysql-connector-python
python-dotenv

# Como usar
pip install -r requirements.txt